requirements.txt
PyGithub
openai
unidiff
python-dotenv